Moved ModalNavigationDrawer to separate file Edit debug menu visible properties(Two taps on logo in ModalNavigationDrawer)
